Change to polling method for IRQs
This is a short-term solution to the problem of interrupt identification.
The Wi-Fi connection creates interrupts, which caused a bug where the
shades would repeatedly toggle. The long-term solution is to wait specifically
for those 2 interrupts instead of waiting for any interrupt. This will
allow support for Wi-Fi an dUSB serial connection